+++ rpc/ethereum/backend/backend.go
@@ -12,9 +12,11 @@ import (
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
 	"github.com/cosmos/cosmos-sdk/server"
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	authtx "github.com/cosmos/cosmos-sdk/x/auth/tx"
 	"github.com/ethereum/go-ethereum/accounts/keystore"
 	"github.com/ethereum/go-ethereum/params"
+	"github.com/ethereum/go-ethereum/rpc"
 	tmrpctypes "github.com/tendermint/tendermint/rpc/core/types"
 
 	"google.golang.org/grpc"
@@ -42,6 +44,9 @@ import (
 // Backend implements the functionality shared within namespaces.
 // Implemented by EVMBackend.
 type Backend interface {
+	// Fee API
+	FeeHistory(blockCount rpc.DecimalOrHex, lastBlock rpc.BlockNumber, rewardPercentiles []float64) (*types.FeeHistoryResult, error)
+
 	// General Ethereum API
 	RPCGasCap() uint64            // global gas cap for eth_call over rpc: DoS protection
 	RPCEVMTimeout() time.Duration // global timeout for eth_call over rpc: DoS protection
@@ -76,9 +81,9 @@ type Backend interface {
 	GetLogs(hash common.Hash) ([][]*ethtypes.Log, error)
 	GetLogsByHeight(height *int64) ([][]*ethtypes.Log, error)
 	GetFilteredBlocks(from int64, to int64, filter [][]filters.BloomIV, filterAddresses bool) ([]int64, error)
-
 	ChainConfig() *params.ChainConfig
 	SetTxDefaults(args evmtypes.TransactionArgs) (evmtypes.TransactionArgs, error)
+	GetEthereumMsgsFromTendermintBlock(block *tmrpctypes.ResultBlock) []*evmtypes.MsgEthereumTx
 }
 
 var _ Backend = (*EVMBackend)(nil)
@@ -355,6 +360,13 @@ func (e *EVMBackend) EthBlockFromTendermint(
 		return nil, err
 	}
 
+	resBlockResult, err := e.clientCtx.Client.BlockResults(ctx, &block.Height)
+	if err != nil {
+		return nil, err
+	}
+
+	txResults := resBlockResult.TxsResults
+
 	for i, txBz := range block.Txs {
 		tx, err := e.clientCtx.TxConfig.TxDecoder()(txBz)
 		if err != nil {
@@ -369,6 +381,12 @@ func (e *EVMBackend) EthBlockFromTendermint(
 			}
 
 			tx := ethMsg.AsTransaction()
+
+			// check tx exists on EVM by cross checking with blockResults
+			if txResults[i].Code != 0 {
+				e.logger.Debug("invalid tx result code", "hash", tx.Hash().Hex())
+				continue
+			}
 
 			if !fullTx {
 				hash := tx.Hash()
@@ -418,20 +436,14 @@ func (e *EVMBackend) EthBlockFromTendermint(
 
 	validatorAddr := common.BytesToAddress(addr)
 
-	gasLimit, err := types.BlockMaxGasFromConsensusParams(ctx, e.clientCtx)
+	gasLimit, err := types.BlockMaxGasFromConsensusParams(ctx, e.clientCtx, block.Height)
 	if err != nil {
 		e.logger.Error("failed to query consensus params", "error", err.Error())
 	}
 
-	resBlockResult, err := e.clientCtx.Client.BlockResults(e.ctx, &block.Height)
-	if err != nil {
-		e.logger.Debug("EthBlockFromTendermint block result not found", "height", block.Height, "error", err.Error())
-		return nil, err
-	}
-
 	gasUsed := uint64(0)
 
-	for _, txsResult := range resBlockResult.TxsResults {
+	for _, txsResult := range txResults {
 		gasUsed += uint64(txsResult.GetGasUsed())
 	}
 
@@ -641,11 +653,12 @@ func (e *EVMBackend) GetCoinbase() (sdk.AccAddress, error) {
 // GetTransactionByHash returns the Ethereum format transaction identified by Ethereum transaction hash
 func (e *EVMBackend) GetTransactionByHash(txHash common.Hash) (*types.RPCTransaction, error) {
 	res, err := e.GetTxByEthHash(txHash)
+	hexTx := txHash.Hex()
 	if err != nil {
 		// try to find tx in mempool
 		txs, err := e.PendingTransactions()
 		if err != nil {
-			e.logger.Debug("tx not found", "hash", txHash.Hex(), "error", err.Error())
+			e.logger.Debug("tx not found", "hash", hexTx, "error", err.Error())
 			return nil, nil
 		}
 
@@ -656,7 +669,7 @@ func (e *EVMBackend) GetTransactionByHash(txHash common.Hash) (*types.RPCTransac
 				continue
 			}
 
-			if msg.Hash == txHash.Hex() {
+			if msg.Hash == hexTx {
 				rpctx, err := types.NewTransactionFromMsg(
 					msg,
 					common.Hash{},
@@ -671,7 +684,7 @@ func (e *EVMBackend) GetTransactionByHash(txHash common.Hash) (*types.RPCTransac
 			}
 		}
 
-		e.logger.Debug("tx not found", "hash", txHash.Hex())
+		e.logger.Debug("tx not found", "hash", hexTx)
 		return nil, nil
 	}
 
@@ -681,23 +694,23 @@ func (e *EVMBackend) GetTransactionByHash(txHash common.Hash) (*types.RPCTransac
 		return nil, nil
 	}
 
-	tx, err := e.clientCtx.TxConfig.TxDecoder()(res.Tx)
-	if err != nil {
-		e.logger.Debug("decoding failed", "error", err.Error())
-		return nil, fmt.Errorf("failed to decode tx: %w", err)
+	var txIndex uint64
+	msgs := e.GetEthereumMsgsFromTendermintBlock(resBlock)
+
+	for i := range msgs {
+		if msgs[i].Hash == hexTx {
+			txIndex = uint64(i)
+			break
+		}
 	}
 
-	msg, err := evmtypes.UnwrapEthereumMsg(&tx)
-	if err != nil {
-		e.logger.Debug("invalid tx", "error", err.Error())
-		return nil, err
-	}
+	msg := msgs[txIndex]
 
 	return types.NewTransactionFromMsg(
 		msg,
 		common.BytesToHash(resBlock.Block.Hash()),
 		uint64(res.Height),
-		uint64(res.Index),
+		txIndex,
 		e.chainID,
 	)
 }
@@ -798,10 +811,10 @@ func (e *EVMBackend) SendTransaction(args evmtypes.TransactionArgs) (common.Hash
 	// NOTE: If error is encountered on the node, the broadcast will not return an error
 	syncCtx := e.clientCtx.WithBroadcastMode(flags.BroadcastSync)
 	rsp, err := syncCtx.BroadcastTx(txBytes)
-	if err != nil || rsp.Code != 0 {
-		if err == nil {
-			err = errors.New(rsp.RawLog)
-		}
+	if rsp != nil && rsp.Code != 0 {
+		err = sdkerrors.ABCIError(rsp.Codespace, rsp.Code, rsp.RawLog)
+	}
+	if err != nil {
 		e.logger.Error("failed to broadcast tx", "error", err.Error())
 		return txHash, err
 	}
@@ -878,6 +891,11 @@ func (e *EVMBackend) RPCTxFeeCap() float64 {
 // RPCFilterCap is the limit for total number of filters that can be created
 func (e *EVMBackend) RPCFilterCap() int32 {
 	return e.cfg.JSONRPC.FilterCap
+}
+
+// RPCFeeHistoryCap is the limit for total number of blocks that can be fetched
+func (e *EVMBackend) RPCFeeHistoryCap() int32 {
+	return e.cfg.JSONRPC.FeeHistoryCap
 }
 
 // RPCMinGasPrice returns the minimum gas price for a transaction obtained from
@@ -999,6 +1017,39 @@ BLOCKS:
 	}
 
 	return matchedBlocks, nil
+}
+
+// GetEthereumMsgsFromTendermintBlock returns all real MsgEthereumTxs from a Tendermint block.
+// It also ensures consistency over the correct txs indexes across RPC endpoints
+func (e *EVMBackend) GetEthereumMsgsFromTendermintBlock(block *tmrpctypes.ResultBlock) []*evmtypes.MsgEthereumTx {
+	var result []*evmtypes.MsgEthereumTx
+
+	for _, tx := range block.Block.Txs {
+		tx, err := e.clientCtx.TxConfig.TxDecoder()(tx)
+		if err != nil {
+			e.logger.Debug("failed to decode transaction in block", "height", block.Block.Height, "error", err.Error())
+			continue
+		}
+
+		for _, msg := range tx.GetMsgs() {
+			ethMsg, ok := msg.(*evmtypes.MsgEthereumTx)
+			if !ok {
+				continue
+			}
+
+			hash := ethMsg.AsTransaction().Hash()
+			// check tx exists on EVM and has the correct block height
+			ethTx, err := e.GetTxByEthHash(hash)
+			if err != nil || ethTx.Height != block.Block.Height {
+				e.logger.Debug("failed to query eth tx hash", "hash", hash.Hex())
+				continue
+			}
+
+			result = append(result, ethMsg)
+		}
+	}
+
+	return result
 }
 
 // checkMatches revised the function from