+++ rpc/ethereum/namespaces/eth/api.go
@@ -22,6 +22,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/cosmos/cosmos-sdk/crypto/keyring"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 
 	"github.com/ethereum/go-ethereum/accounts/keystore"
@@ -56,7 +57,7 @@ func NewPublicAPI(
 	backend backend.Backend,
 	nonceLock *rpctypes.AddrLocker,
 ) *PublicAPI {
-	epoch, err := ethermint.ParseChainID(clientCtx.ChainID)
+	eip155ChainID, err := ethermint.ParseChainID(clientCtx.ChainID)
 	if err != nil {
 		panic(err)
 	}
@@ -80,13 +81,18 @@ func NewPublicAPI(
 
 	// The signer used by the API should always be the 'latest' known one because we expect
 	// signers to be backwards-compatible with old transactions.
-	signer := ethtypes.LatestSigner(backend.ChainConfig())
+	cfg := backend.ChainConfig()
+	if cfg == nil {
+		cfg = evmtypes.DefaultChainConfig().EthereumConfig(eip155ChainID)
+	}
+
+	signer := ethtypes.LatestSigner(cfg)
 
 	api := &PublicAPI{
 		ctx:          context.Background(),
 		clientCtx:    clientCtx,
 		queryClient:  rpctypes.NewQueryClient(clientCtx),
-		chainIDEpoch: epoch,
+		chainIDEpoch: eip155ChainID,
 		logger:       logger.With("client", "json-rpc"),
 		backend:      backend,
 		nonceLock:    nonceLock,
@@ -212,8 +218,7 @@ func (e *PublicAPI) MaxPriorityFeePerGas() (*hexutil.Big, error) {
 
 func (e *PublicAPI) FeeHistory(blockCount rpc.DecimalOrHex, lastBlock rpc.BlockNumber, rewardPercentiles []float64) (*rpctypes.FeeHistoryResult, error) {
 	e.logger.Debug("eth_feeHistory")
-
-	return nil, fmt.Errorf("eth_feeHistory not implemented")
+	return e.backend.FeeHistory(blockCount, lastBlock, rewardPercentiles)
 }
 
 // Accounts returns the list of accounts available to this node.
@@ -315,7 +320,8 @@ func (e *PublicAPI) GetBlockTransactionCountByHash(hash common.Hash) *hexutil.Ui
 		return nil
 	}
 
-	n := hexutil.Uint(len(resBlock.Block.Txs))
+	ethMsgs := e.backend.GetEthereumMsgsFromTendermintBlock(resBlock)
+	n := hexutil.Uint(len(ethMsgs))
 	return &n
 }
 
@@ -333,7 +339,8 @@ func (e *PublicAPI) GetBlockTransactionCountByNumber(blockNum rpctypes.BlockNumb
 		return nil
 	}
 
-	n := hexutil.Uint(len(resBlock.Block.Txs))
+	ethMsgs := e.backend.GetEthereumMsgsFromTendermintBlock(resBlock)
+	n := hexutil.Uint(len(ethMsgs))
 	return &n
 }
 
@@ -473,10 +480,10 @@ func (e *PublicAPI) SendRawTransaction(data hexutil.Bytes) (common.Hash, error) 
 
 	syncCtx := e.clientCtx.WithBroadcastMode(flags.BroadcastSync)
 	rsp, err := syncCtx.BroadcastTx(txBytes)
-	if err != nil || rsp.Code != 0 {
-		if err == nil {
-			err = errors.New(rsp.RawLog)
-		}
+	if rsp != nil && rsp.Code != 0 {
+		err = sdkerrors.ABCIError(rsp.Codespace, rsp.Code, rsp.RawLog)
+	}
+	if err != nil {
 		e.logger.Error("failed to broadcast tx", "error", err.Error())
 		return txHash, err
 	}
@@ -673,23 +680,13 @@ func (e *PublicAPI) GetTransactionByBlockHashAndIndex(hash common.Hash, idx hexu
 	}
 
 	i := int(idx)
-	if i >= len(resBlock.Block.Txs) {
+	ethMsgs := e.backend.GetEthereumMsgsFromTendermintBlock(resBlock)
+	if i >= len(ethMsgs) {
 		e.logger.Debug("block txs index out of bound", "index", i)
 		return nil, nil
 	}
 
-	txBz := resBlock.Block.Txs[i]
-	tx, err := e.clientCtx.TxConfig.TxDecoder()(txBz)
-	if err != nil {
-		e.logger.Debug("decoding failed", "error", err.Error())
-		return nil, fmt.Errorf("failed to decode tx: %w", err)
-	}
-
-	msg, err := evmtypes.UnwrapEthereumMsg(&tx)
-	if err != nil {
-		e.logger.Debug("invalid tx", "error", err.Error())
-		return nil, err
-	}
+	msg := ethMsgs[i]
 
 	baseFee, err := e.backend.BaseFee(resBlock.Block.Height)
 	if err != nil {
@@ -721,23 +718,13 @@ func (e *PublicAPI) GetTransactionByBlockNumberAndIndex(blockNum rpctypes.BlockN
 	}
 
 	i := int(idx)
-	if i >= len(resBlock.Block.Txs) {
+	ethMsgs := e.backend.GetEthereumMsgsFromTendermintBlock(resBlock)
+	if i >= len(ethMsgs) {
 		e.logger.Debug("block txs index out of bound", "index", i)
 		return nil, nil
 	}
 
-	txBz := resBlock.Block.Txs[i]
-	tx, err := e.clientCtx.TxConfig.TxDecoder()(txBz)
-	if err != nil {
-		e.logger.Debug("decoding failed", "error", err.Error())
-		return nil, fmt.Errorf("failed to decode tx: %w", err)
-	}
-
-	msg, err := evmtypes.UnwrapEthereumMsg(&tx)
-	if err != nil {
-		e.logger.Debug("invalid tx", "error", err.Error())
-		return nil, err
-	}
+	msg := ethMsgs[i]
 
 	return rpctypes.NewTransactionFromMsg(
 		msg,
@@ -750,11 +737,12 @@ func (e *PublicAPI) GetTransactionByBlockNumberAndIndex(blockNum rpctypes.BlockN
 
 // GetTransactionReceipt returns the transaction receipt identified by hash.
 func (e *PublicAPI) GetTransactionReceipt(hash common.Hash) (map[string]interface{}, error) {
-	e.logger.Debug("eth_getTransactionReceipt", "hash", hash.Hex())
+	hexTx := hash.Hex()
+	e.logger.Debug("eth_getTransactionReceipt", "hash", hexTx)
 
 	res, err := e.backend.GetTxByEthHash(hash)
 	if err != nil {
-		e.logger.Debug("tx not found", "hash", hash.Hex(), "error", err.Error())
+		e.logger.Debug("tx not found", "hash", hexTx, "error", err.Error())
 		return nil, nil
 	}
 
@@ -808,7 +796,17 @@ func (e *PublicAPI) GetTransactionReceipt(hash common.Hash) (map[string]interfac
 
 	logs, err := e.backend.GetTransactionLogs(hash)
 	if err != nil {
-		e.logger.Debug("logs not found", "hash", hash.Hex(), "error", err.Error())
+		e.logger.Debug("logs not found", "hash", hexTx, "error", err.Error())
+	}
+
+	// get eth index based on block's txs
+	var txIndex uint64
+	msgs := e.backend.GetEthereumMsgsFromTendermintBlock(resBlock)
+	for i := range msgs {
+		if msgs[i].Hash == hexTx {
+			txIndex = uint64(i)
+			break
+		}
 	}
 
 	receipt := map[string]interface{}{
@@ -829,7 +827,7 @@ func (e *PublicAPI) GetTransactionReceipt(hash common.Hash) (map[string]interfac
 		// transaction corresponding to this receipt.
 		"blockHash":        common.BytesToHash(resBlock.Block.Header.Hash()).Hex(),
 		"blockNumber":      hexutil.Uint64(res.Height),
-		"transactionIndex": hexutil.Uint64(res.Index),
+		"transactionIndex": hexutil.Uint64(txIndex),
 
 		// sender and receiver (contract or EOA) addreses
 		"from": from,
@@ -848,7 +846,7 @@ func (e *PublicAPI) GetTransactionReceipt(hash common.Hash) (map[string]interfac
 	return receipt, nil
 }
 
-// PendingTransactions returns the transactions that are in the transaction pool
+// GetPendingTransactions returns the transactions that are in the transaction pool
 // and have a from address that is one of the accounts this node manages.
 func (e *PublicAPI) GetPendingTransactions() ([]*rpctypes.RPCTransaction, error) {
 	e.logger.Debug("eth_getPendingTransactions")